+++ src/Symfony/Bridge/Twig/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 -----
 
  * added TwigFlattener which adds twig files into the FlattenException
+ * added PHP and Twig syntax highlighters
 
 2.5.0
 -----